refactor project to prepare vs2022 extension
