+++ src/VSExtension/ReswPlusVSIX.2019/Assets/Logo.png
