+++ src/VSExtension/ReswPlusVSIX.2019/Assets/ReswPlusGenerateIco.png
