+++ src/VSExtension/ReswPlusVSIX.2019/Assets/ReswPlusPackage.ico
