+++ src/VSExtension/ReswPlusVSIX.2019/Key.snk
