+++ src/VSExtension/ReswPlusVSIX.2019/Resources/ContextMenuPackage.ico
