+++ src/VSExtension/ReswPlusVSIX.2019/Resources/menuIcon.bmp
