+++ src/VSExtension/ReswPlusVSIX.Shared/Converters/AndroidXMLConverter.cs
