+++ src/VSExtension/ReswPlusVSIX.Shared/ResourceInfo/ProjectInfo.cs
