+++ src/VSExtension/ReswPlusVSIX.Shared/ResourceInfo/ResourceFileInfoBuilder.cs
