+++ src/VSExtension/ReswPlusVSIX.Shared/ReswPlusVSIX.Shared.projitems
@@ -1,0 +1,30 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<Project xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
+  <PropertyGroup>
+    <MSBuildAllProjects Condition="'$(MSBuildVersion)' == '' Or '$(MSBuildVersion)' &lt; '16.0'">$(MSBuildAllProjects);$(MSBuildThisFileFullPath)</MSBuildAllProjects>
+    <HasSharedItems>true</HasSharedItems>
+    <SharedGUID>0ad6a890-4d11-463d-b547-0a227317dd22</SharedGUID>
+  </PropertyGroup>
+  <PropertyGroup Label="Configuration">
+    <Import_RootNamespace>ReswPlusVSIX.Shared</Import_RootNamespace>
+  </PropertyGroup>
+  <ItemGroup>
+    <Compile Include="$(MSBuildThisFileDirectory)ContextMenu.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Converters\AndroidXMLConverter.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)ResourceInfo\ProjectInfo.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)ResourceInfo\ResourceFileInfoBuilder.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)ReswPlusPackage.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)SingleFileGenerators\ReswPlusAdvancedGenerator.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)SingleFileGenerators\ReswPlusGenerator.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)SingleFileGenerators\ReswPlusGeneratorBase.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Utils\CodeGeneratorRegistrationAttribute.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Utils\NugetHelper.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Utils\ProjectExtension.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)VSUIIntegration.cs" />
+  </ItemGroup>
+  <ItemGroup>
+    <None Include="$(MSBuildThisFileDirectory)source.extension.vsixmanifest">
+      <SubType>Designer</SubType>
+    </None>
+  </ItemGroup>
+</Project>
