+++ src/VSExtension/ReswPlusVSIX.Shared/Utils/NugetHelper.cs
