+++ src/VSExtension/ReswPlusVSIX.Shared/Utils/ProjectExtension.cs
