+++ src/VSExtension/ReswPlusVSIX.Shared/VSUIIntegration.cs
