--- src/VSExtension/ReswPlusVSIX/packages.config
@@ -1,40 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="EnvDTE" version="8.0.2" targetFramework="net46" />
-  <package id="EnvDTE100" version="10.0.3" targetFramework="net46" />
-  <package id="EnvDTE80" version="8.0.3" targetFramework="net46" />
-  <package id="EnvDTE90" version="9.0.3" targetFramework="net46" />
-  <package id="EnvDTE90a" version="9.0.3" targetFramework="net46" />
-  <package id="Microsoft.VisualStudio.ComponentModelHost" version="15.8.525" targetFramework="net462" />
-  <package id="Microsoft.VisualStudio.CoreUtility" version="15.0.26606" targetFramework="net462" />
-  <package id="Microsoft.VisualStudio.Designer.Interfaces" version="1.1.4323" targetFramework="net462" />
-  <package id="Microsoft.VisualStudio.Imaging" version="15.0.26606" targetFramework="net462" />
-  <package id="Microsoft.VisualStudio.Imaging.Interop.14.0.DesignTime" version="14.3.26930" targetFramework="net462" />
-  <package id="Microsoft.VisualStudio.OLE.Interop" version="7.10.6071" targetFramework="net46" />
-  <package id="Microsoft.VisualStudio.SDK.Analyzers" version="15.8.36" targetFramework="net462" />
-  <package id="Microsoft.VisualStudio.SDK.EmbedInteropTypes" version="15.0.27" targetFramework="net462" />
-  <package id="Microsoft.VisualStudio.Shell.15.0" version="15.0.26228" targetFramework="net462" />
-  <package id="Microsoft.VisualStudio.Shell.Framework" version="15.0.26228" targetFramework="net462" />
-  <package id="Microsoft.VisualStudio.Shell.Interop" version="7.10.6072" targetFramework="net462" />
-  <package id="Microsoft.VisualStudio.Shell.Interop.10.0" version="10.0.30320" targetFramework="net462" />
-  <package id="Microsoft.VisualStudio.Shell.Interop.11.0" version="11.0.61031" targetFramework="net462" />
-  <package id="Microsoft.VisualStudio.Shell.Interop.12.0" version="12.0.30111" targetFramework="net462" />
-  <package id="Microsoft.VisualStudio.Shell.Interop.14.0.DesignTime" version="14.3.26929" targetFramework="net462" />
-  <package id="Microsoft.VisualStudio.Shell.Interop.8.0" version="8.0.50728" targetFramework="net462" />
-  <package id="Microsoft.VisualStudio.Shell.Interop.9.0" version="9.0.30730" targetFramework="net462" />
-  <package id="Microsoft.VisualStudio.TextManager.Interop" version="7.10.6071" targetFramework="net462" />
-  <package id="Microsoft.VisualStudio.TextManager.Interop.8.0" version="8.0.50728" targetFramework="net462" />
-  <package id="Microsoft.VisualStudio.Threading" version="15.3.83" targetFramework="net462" />
-  <package id="Microsoft.VisualStudio.Threading.Analyzers" version="15.8.168" targetFramework="net462" />
-  <package id="Microsoft.VisualStudio.Utilities" version="15.0.26607" targetFramework="net462" />
-  <package id="Microsoft.VisualStudio.Validation" version="15.3.58" targetFramework="net462" />
-  <package id="Microsoft.VSSDK.BuildTools" version="15.8.3252" targetFramework="net462" developmentDependency="true" />
-  <package id="NuGet.VisualStudio" version="4.8.2" targetFramework="net462" />
-  <package id="stdole" version="7.0.3303" targetFramework="net462" />
-  <package id="System.ValueTuple" version="4.5.0" targetFramework="net462" />
-  <package id="VSLangProj" version="7.0.3301" targetFramework="net462" />
-  <package id="VSSDK.DTE" version="7.0.4" targetFramework="net462" />
-  <package id="VSSDK.IDE" version="7.0.4" targetFramework="net462" />
-  <package id="VSSDK.IDE.12" version="12.0.4" targetFramework="net462" />
-  <package id="VSSDK.TemplateWizardInterface" version="12.0.4" targetFramework="net462" />
-</packages>
